+++ types/env.d.ts
@@ -8,6 +8,12 @@ declare global {
       NODE_ENV: 'development' | 'test' | 'production'
 
       HTTP_PORT: string // numeric
+
+      DB_HOST?: string
+      DB_PORT?: string // numeric
+      DB_NAME: string
+      DB_USER: string
+      DB_PASS: string
     }
   }
 }