+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog].
 
 - `CHANGELOG.md` file ([6])
 - Swagger UI for API documentation ([7])
+- PostgreSQL database integration via Drizzle ORM ([2])
 
 ### Changed
 
@@ -21,7 +22,8 @@ The format is based on [Keep a Changelog].
 
 ### Internal
 
-- `@nestjs/swagger` package for Swagger integration (#7)
+- Added `@nestjs/swagger` package ([7])
+- Added `drizzle-orm`, `pg`, `@types/pg`, `drizzle-kit` packages ([2])
 
 <!-- Links -->
 
@@ -31,6 +33,7 @@ The format is based on [Keep a Changelog].
 
 [7]: https://github.com/i19n-io/backend/pull/7
 [6]: https://github.com/i19n-io/backend/pull/6
+[2]: https://github.com/i19n-io/backend/pull/2
 
 <!-- Links to versions -->